forgot that this endpoint was testonly
